Added for FINANCIAL ACCOUNTING YEAR CONFIGURATION

+++ application/projects/src/main/java/com/idg/idgcore/coe/domain/entity/financialAccountingYear/FinancialAccountingYearEntity.java
@@ -19,7 +19,7 @@ public class FinancialAccountingYearEntity extends AbstractAuditableDomainEntity
 {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
-    private Integer financialAccountingYearId;
+    private Long financialAccountingYearId;
     private String bankCode;
     private String branchCode;
     private Date startDate;
