+++ application/projects/src/main/java/com/idg/idgcore/coe/dto/financialAccountingYear/FinancialAccountingYearPeriodicCodeDTO.java
@@ -10,7 +10,7 @@ import java.util.*;
 @NoArgsConstructor
 public class FinancialAccountingYearPeriodicCodeDTO {
 
-    private Integer finAccYearPeriodCodesId;
+    private Long finAccYearPeriodCodesId;
     private String periodCode;
     private Date startDateAccountingPeriod;
     private Date endDateAccountingPeriod;
